move widget reducer to its own file
